fix(completions): ignore aliases and functions named usage (#300)
Resolves https://github.com/jdx/mise/discussions/5358.

This PR replaces `usage` with `command usage`, to ignore aliases and
functions named `usage`.

Also, for consistency, use `command usage` instead of `command -v usage`
to check its existence.
In this way, if aliases or functions named `usage` exist but `usage` is
not installed, the completions are skipped as expected.

I don't think this is a case, but something like `alias usage=usage
--default-flag` will be ignored by this PR.