fix(completions): ignore aliases and functions named usage (2nd attempt) (#304)
Reverts jdx/usage#301. (Reapplies #300.)

Replaces `command -v usage` with `type -p usage` to ignore aliases and
functions.
In the 1st attempt, I used `command usage`, but since `usage` (or `usage
--help`) exits with an exit code `2`, it didn't work as an existence
check.

I apologise for the trouble. Thank you for your help as always.

---

Resolves https://github.com/jdx/mise/discussions/5358.

This PR replaces `usage` with `command usage`, to ignore aliases and
functions named `usage`.

Also, for consistency, use `command usage` instead of `command -v usage`
to check its existence.
In this way, if aliases or functions named `usage` exist but `usage` is
not installed, the completions are skipped as expected.

I don't think this is a case, but something like `alias usage=usage
--default-flag` will be ignored by this PR.