fix: add fallback for shell
In https://github.com/jdx/usage/pull/341, a "known" shell type was made
mandatory for the complete word command. Since `shell` is an optional
argument some kind of valid fallback should be supported.

To fix this `shell` is no longer optional and is provided with a default
value of "bash". Added a unit test to validate behavior going forward.